feat: Add JFrog Artifactory upload action
This commit adds a new action, "[Ledger Security] Upload to JFrog Artifactory", which is a wrapper for the JFrog Artifactory upload functionality. It allows users to specify the path to the file to upload and the name of the Artifactory repository to upload to. The action uses the jfrog/setup-jfrog-cli@v4 action to set up the JFrog CLI and then runs the "jfrog rt u" command to perform the upload.