feat: Update JFrog login action.yaml to use steps.setup-jfrog-cli.outputs
This commit updates the JFrog login action.yaml file to use the `steps.setup-jfrog-cli.outputs` syntax instead of `setup-jfrog-cli.outputs`. This change ensures that the JFrog OIDC token and username are correctly retrieved from the Setup JFrog CLI step and passed as outputs. By making this update, users will be able to access and use these values in subsequent steps or workflows.